Day one:
Day one week one:
I have learned basic code such as the use of "commit." I have also learned the process of forming branches which are used as pointers to a specific commit. We start with the use of 'git checkout -b [name of branch].' This creates a new branch which we can then separate from the master branch with a commit, followed by 'git checkout master,' followed by another commit.

 I then learned the process of merging. This is done with the same method as creating the branches, followed by the use of 'Git merge.' If we wanted to merge Master with [name of branch] we would use "git merge [name of other branch],' as long as the master branch is selected ect.

The final thing we learned about was rebase. A rebase is the task of taking a set of commits and copying them to another location. The process of this is similar to that of merging in the previous task, but instead of using 'git merge' we use 'git rebase master.' In this case  we are wanting to copy the [name of other branch] to the master, so we would want the other branch to be selected not the master branch.